authentication

+++ Blogify/app/Http/Controllers/BlogController.php
@@ -16,8 +16,10 @@ class BlogController extends Controller
     {
         $validated = $request->validate([
             'title' => 'required|min:1|max:150',
-            'content' => 'required|min:1|max:2000'
+            'content' => 'required|min:1|max:2000',
         ]);
+
+        $validated['user_id'] = auth()->id();
 
         Blog::create($validated);
 
@@ -26,25 +28,39 @@ class BlogController extends Controller
 
     public function destroy(Blog $blog)
     {
+        if (auth()->id() !== $blog->user_id) {
+            abort(404);
+        }
+
         $blog->delete();
         return redirect()->route('dashboard')->with('successful', 'Blog deleted successfully');
     }
 
     public function edit(Blog $blog)
     {
+        if (auth()->id() !== $blog->user_id) {
+            abort(404);
+        }
+
         $editing = true;
         return view('blogs.show', compact('blog', 'editing'));
     }
 
     public function update(Request $request, Blog $blog)
     {
+        if (auth()->id() !== $blog->user_id) {
+            abort(404);
+        }
+
         $validated = $request->validate([
             'title' => 'required|min:1|max:150',
-            'content' => 'required|min:1|max:2000'
+            'content' => 'required|min:1|max:2000',
         ]);
 
         $blog->update($validated);
 
-        return redirect()->route('blogs.show', $blog->id)->with('successful', 'Blog updated successfully');
+        return redirect()
+            ->route('blogs.show', $blog->id)
+            ->with('successful', 'Blog updated successfully');
     }
 }
